requirements.txt
async-timeout==4.0.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.1
websockets==12.0
